# Improve performance of async deserialization