--- a/samples/.editorconfig
+++ b/samples/.editorconfig
@@ -52,3 +52,6 @@ dotnet_diagnostic.CS1591.severity = silent
 
 # CA1822: Mark members as static
 dotnet_diagnostic.CA1822.severity = silent
+
+# CA1062: Validate arguments of public methods
+dotnet_diagnostic.CA1062.severity = silent