--- a/src/Nerdbank.MessagePack/Converters/ArrayConverter`1.cs
+++ b/src/Nerdbank.MessagePack/Converters/ArrayConverter`1.cs
@@ -99,7 +99,7 @@ internal class ArrayConverter<TElement>(MessagePackConverter<TElement> elementCo
 		bool success;
 		while (streamingReader.TryReadNil(out success).NeedsMoreBytes())
 		{
-			streamingReader = new(await streamingReader.ReadMoreBytesAsync());
+			streamingReader = new(await streamingReader.FetchMoreBytesAsync());
 		}
 
 		if (success)
@@ -114,7 +114,7 @@ internal class ArrayConverter<TElement>(MessagePackConverter<TElement> elementCo
 			int count;
 			while (streamingReader.TryReadArrayHeader(out count).NeedsMoreBytes())
 			{
-				streamingReader = new(await streamingReader.ReadMoreBytesAsync());
+				streamingReader = new(await streamingReader.FetchMoreBytesAsync());
 			}
 
 			reader.ReturnReader(ref streamingReader);