--- a/src/Nerdbank.MessagePack/Converters/ObjectArrayWithNonDefaultCtorConverter`2.cs
+++ b/src/Nerdbank.MessagePack/Converters/ObjectArrayWithNonDefaultCtorConverter`2.cs
@@ -85,7 +85,7 @@ internal class ObjectArrayWithNonDefaultCtorConverter<TDeclaringType, TArgumentS
 		bool success;
 		while (streamingReader.TryReadNil(out success).NeedsMoreBytes())
 		{
-			streamingReader = new(await streamingReader.ReplenishBufferAsync());
+			streamingReader = new(await streamingReader.ReadMoreBytes());
 		}
 
 		if (success)
@@ -100,7 +100,7 @@ internal class ObjectArrayWithNonDefaultCtorConverter<TDeclaringType, TArgumentS
 		MessagePackType peekType;
 		while (streamingReader.TryPeekNextMessagePackType(out peekType).NeedsMoreBytes())
 		{
-			streamingReader = new(await streamingReader.ReplenishBufferAsync());
+			streamingReader = new(await streamingReader.ReadMoreBytes());
 		}
 
 		if (peekType == MessagePackType.Map)
@@ -108,7 +108,7 @@ internal class ObjectArrayWithNonDefaultCtorConverter<TDeclaringType, TArgumentS
 			int mapEntries;
 			while (streamingReader.TryReadMapHeader(out mapEntries).NeedsMoreBytes())
 			{
-				streamingReader = new(await streamingReader.ReplenishBufferAsync());
+				streamingReader = new(await streamingReader.ReadMoreBytes());
 			}
 
 			// We're going to read in bursts. Anything we happen to get in one buffer, we'll read synchronously regardless of whether the property is async.
@@ -171,7 +171,7 @@ internal class ObjectArrayWithNonDefaultCtorConverter<TDeclaringType, TArgumentS
 			int arrayLength;
 			while (streamingReader.TryReadArrayHeader(out arrayLength).NeedsMoreBytes())
 			{
-				streamingReader = new(await streamingReader.ReplenishBufferAsync());
+				streamingReader = new(await streamingReader.ReadMoreBytes());
 			}
 
 			reader.ReturnReader(ref streamingReader);