… src/Nerdbank.MessagePack.Analyzers/AnalyzerReleases.Unshipped.md
… src/Nerdbank.MessagePack/DecodeResultExtensions.cs
… src/Nerdbank.MessagePack/MessagePackAsyncReader.cs
… src/Nerdbank.MessagePack/MessagePackConverter`1.cs
… src/Nerdbank.MessagePack/MessagePackPrimitives.Writers.cs
… src/Nerdbank.MessagePack/MessagePackReader.Integers.cs
… src/Nerdbank.MessagePack/MessagePackReader.cs
… src/Nerdbank.MessagePack/MessagePackSerializer.cs
… src/Nerdbank.MessagePack/MessagePackStreamingReader.Integers.cs
… src/Nerdbank.MessagePack/MessagePackStreamingReader.Integers.tt
… src/Nerdbank.MessagePack/MessagePackStreamingReader.cs
… src/Nerdbank.MessagePack/Nerdbank.MessagePack.csproj
… src/Nerdbank.MessagePack/StandardVisitor.cs
… src/Nerdbank.MessagePack/net8.0/PublicAPI.Unshipped.txt
… src/Nerdbank.MessagePack/netstandard2.0/PublicAPI.Unshipped.txt
… test/Nerdbank.MessagePack.Tests/MessagePackStreamingReaderTests.cs